Splits MeetupServer into its own module
Also flattens the server's internal state into a single struct, in
preparation for a public trait for publisher data stores.
